Add playRound and playGame functions
These functions allow the user to play Rock, Paper, Scissors
in the console.

The playRound function compares the return value of getHumanChoice()
and getComputerChoice(). A message is printed in the console informing
the user whether they won, lost, or tied with the computer in that
round. Whoever wins that round has their score incremented by 1.

The playGame function runs the playRound function five times and prints
the scores of both players after the five rounds are played. The scores
are reset at the beginning of each game.

The capitalizeFirstLetter function was also added to capitalize
the first letter "rock", "paper", and "scissors" in the console.log
messages.